// src/modules/scope/enums/scope-code.enum.ts
export enum ScopeCodeEnum {
  CATEGORY_MANAGEMENT = 'CATEGORY_MANAGEMENT',
  ADMIN_MANAGEMENT = 'ADMIN_MANAGEMENT',
  ROLES_MANAGEMENT = 'ROLES_MANAGEMENT',
  TRAINING_SESSION_MANAGEMENT = 'TRAINING_SESSION_MANAGEMENT',
  PRESENCE_MANAGEMENT = 'PRESENCE_MANAGEMENT',
  TRAINER_MANAGEMENT = 'TRAINER_MANAGEMENT',
  PLAYER_MANAGEMENT = 'PLAYER_MANAGEMENT',
  TEAMS_MANAGEMENT = 'TEAMS_MANAGEMENT',
}